refactor: extract ffmpeg plan builder from adapter
